+++ packages/checkbox/src/Component.tsx
@@ -63,6 +63,12 @@ export type CheckboxProps = Omit<NativeProps, 'size' | 'onChange' | 'enterKeyHin
     size?: 20 | 24;
 
     /**
+     * Компактный текст подписи Только для `size=20`
+     * @default false
+     */
+    compact?: boolean;
+
+    /**
      * Доп. класс чекбокса
      */
     boxClassName?: string;
@@ -128,8 +134,15 @@ export type CheckboxProps = Omit<NativeProps, 'size' | 'onChange' | 'enterKeyHin
     error?: ReactNode | boolean;
 
     /**
+     * Положение чекбокса относительно контента
+     * @default start
+     */
+    controlPosition?: 'start' | 'end';
+
+    /**
      * Позиция чекбокса относительно контента
      * @default 'before'
+     * @deprecated Используйте prop `controlPosition`
      */
     position?: 'before' | 'after';
 
@@ -169,6 +182,7 @@ export const Checkbox = forwardRef<HTMLLabelElement, CheckboxProps>(
             hintClassName,
             errorClassName,
             addonsClassName,
+            controlPosition = 'start',
             position = 'before',
             align = 'start',
             addons,
@@ -184,6 +198,7 @@ export const Checkbox = forwardRef<HTMLLabelElement, CheckboxProps>(
             error,
             inputRef,
             colors = 'default',
+            compact = false,
             ...restProps
         },
         ref,
@@ -201,6 +216,8 @@ export const Checkbox = forwardRef<HTMLLabelElement, CheckboxProps>(
         const errorMessage = typeof error === 'boolean' ? '' : error;
 
         const colorStyle = colorStyles[colors];
+        const isControlAtEnd =
+            (controlPosition ?? (position === 'after' ? 'end' : 'start')) === 'end';
 
         const renderCheckmark = () => (
             <span className={cn(styles.box, colorStyle.box, boxClassName)}>
@@ -232,7 +249,8 @@ export const Checkbox = forwardRef<HTMLLabelElement, CheckboxProps>(
                         [colorStyle.indeterminate]: indeterminate,
                         [styles.focused]: focused,
                         [styles.block]: block,
-                        [styles['position-after']]: position === 'after',
+                        [styles['position-after']]: isControlAtEnd,
+                        [styles.compact]: compact && size === 20,
                     },
                 )}
                 ref={mergeRefs([labelRef, ref, labelProps?.ref as Ref<HTMLLabelElement>])}
@@ -250,7 +268,7 @@ export const Checkbox = forwardRef<HTMLLabelElement, CheckboxProps>(
                     />
                 )}
 
-                {position === 'before' && renderCheckmark()}
+                {!isControlAtEnd && renderCheckmark()}
 
                 {(label || hint || errorMessage) && (
                     <span className={cn(styles.content, contentClassName)}>
@@ -284,7 +302,7 @@ export const Checkbox = forwardRef<HTMLLabelElement, CheckboxProps>(
                     </span>
                 )}
 
-                {position === 'after' && renderCheckmark()}
+                {isControlAtEnd && renderCheckmark()}
 
                 {addons && (
                     // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions